+++ packages/jest-resolve/src/__tests__/resolve.test.js
@@ -134,8 +134,9 @@ describe('resolveModule', () => {
     const resolver = new Resolver(moduleMap, {
       extensions: ['.js'],
     });
-    const src = require.resolve(
-      '../../src/__mocks__/bar/node_modules/foo/index.js',
+    const src = path.join(
+      path.resolve(__dirname, '../../src/__mocks__/bar/node_modules/'),
+      'foo/index.js',
     );
     const resolved = resolver.resolveModule(src, 'dep');
     expect(resolved).toBe(
@@ -209,16 +210,13 @@ describe('Resolver.getModulePaths() -> nodeModulesPaths()', () => {
       mocks: [],
     });
 
-    // Mocking realpath to function the old way, where it just looks at
-    // pathstrings instead of actually trying to access the physical directory.
-    // This test suite won't work otherwise, since we cannot make assumptions
-    // about the test environment when it comes to absolute paths.
-    jest.doMock('realpath-native', () => ({
-      sync: dirInput => dirInput,
-    }));
+    jest
+      .spyOn(fs, 'lstatSync')
+      .mockImplementation(() => ({isSymbolicLink: () => false}));
   });
 
   afterAll(() => {
+    jest.restoreAllMocks();
     jest.resetModules();
     jest.dontMock('path');
   });